// src/app/admin/pages/suppliers/interfaces/AdminSuppliersInterface.ts
export interface IAdminSupplierInputs {
    suppliername: string
    contactname: string
    personaddress: string
    emailaddress: string
    phone1: string
    phone2: string
    taxNo: string
    cno: string
    maxcredit: number
    maxtime: number
    intbalance: number
    accountno: string
}